backend/requirements.txt
fastapi
asyncpg
pydantic
uvicorn
PyJWT
passlib
datetime
pydantic[email]